Move new type SignedTransactionWithStatus to public package.
