+++ core/src/main/kotlin/net/corda/core/node/services/TransactionStorage.kt
@@ -2,9 +2,11 @@ package net.corda.core.node.services
 
 import net.corda.core.DoNotImplement
 import net.corda.core.concurrent.CordaFuture
+import net.corda.core.contracts.NamedByHash
 import net.corda.core.crypto.SecureHash
-import net.corda.core.internal.SignedTransactionWithStatus
+import net.corda.core.flows.TransactionStatus
 import net.corda.core.messaging.DataFeed
+import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.SignedTransaction
 import rx.Observable
 
@@ -38,4 +40,13 @@ interface TransactionStorage {
      * Returns a future that completes with the transaction corresponding to [id] once it has been committed
      */
     fun trackTransaction(id: SecureHash): CordaFuture<SignedTransaction>
+}
+
+@CordaSerializable
+data class SignedTransactionWithStatus(
+        val stx: SignedTransaction,
+        val status: TransactionStatus
+) : NamedByHash {
+    override val id: SecureHash
+        get() = stx.id
 }
